Rename quick start dir
