+++ examples/a_quick_start/simple_ocr.py
