+++ examples/a_quick_start/summarize_struct.py
