+++ pyproject.toml
@@ -24,7 +24,8 @@ dependencies = [
 
 
 [tool.uv.sources]
-pipelex = { git = "https://github.com/Pipelex/pipelex.git", branch = "feature/Chicago" }
+# pipelex = { git = "https://github.com/Pipelex/pipelex.git", branch = "feature/Chicago" }
+pipelex = { git = "https://github.com/Pipelex/pipelex.git", rev = "fdcc656f5a560a30a6dad3c50b9c6c683416013a" }
 
 
 [project.optional-dependencies]
